Task #24, Updated groupId as per maven conventions
