+++ src/main/java/io/github/handofgod94/Badge.java
@@ -1,4 +1,4 @@
-package org.gahan;
+package io.github.handofgod94;
 
 import java.io.IOException;
 
@@ -14,6 +14,15 @@ public class Badge {
   private float badgeLabelWidth = 0.0f;
   private float badgeValueWidth = 0.0f;
 
+  /**
+   * Enum constants for coverage category.
+   * Jacoco CSV Report generates results of various categories
+   * such as branch missed/covered, instructions missed/covered.
+   * This enum maps each of those categories.
+   *
+   * <p>It includes:
+   * <code>INSTRUCTION, BRANCH, LINE, COMPLEXITY, METHOD</code>
+   */
   public enum CoverageCategory {
     INSTRUCTION,
     BRANCH,
