+++ src/main/java/io/github/handofgod94/BadgeUtility.java
@@ -1,17 +1,17 @@
-package org.gahan;
+package io.github.handofgod94;
 
 import com.opencsv.CSVReader;
 import com.opencsv.CSVReaderBuilder;
+import io.github.handofgod94.format.Formatter;
+import io.github.handofgod94.format.JpegFormatter;
+import io.github.handofgod94.format.PngFormatter;
+import io.github.handofgod94.format.SvgFormatter;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import org.apache.batik.transcoder.TranscoderException;
 import org.apache.pdfbox.pdmodel.font.PDFont;
 import org.apache.pdfbox.pdmodel.font.PDMMType1Font;
-import org.gahan.format.Formatter;
-import org.gahan.format.JpegFormatter;
-import org.gahan.format.PngFormatter;
-import org.gahan.format.SvgFormatter;
 
 
 /**
@@ -55,11 +55,11 @@ public class BadgeUtility {
    * Calculate coverage %age from csv report generated by jacoco.
    *
    * @param csvPath Absolute path of csv report.
-   * @param  category enum representing the coverage category
+   * @param category enum representing the coverage category
    * @return integer value containing total coverage
    * @throws IOException Unable to read jacoco csv report from config location
    * @throws NumberFormatException Unable to parse %ages from jacoco reports
-   * @see org.gahan.Badge.CoverageCategory
+   * @see io.github.handofgod94.Badge.CoverageCategory
    */
   public static int calculateCoverage(String csvPath, Badge.CoverageCategory category)
       throws NumberFormatException, IOException {
@@ -123,9 +123,10 @@ public class BadgeUtility {
    *     <td>BRIGHTGREEN</td>
    *   </tr>
    * </table>
+   *
    * @param coverage coverage percentage rounded off to nearest integer
    * @return An Enum for the given range
-   * @see BadgeColors
+   * @see io.github.handofgod94.BadgeColors
    */
   public static BadgeColors getColorFromRange(int coverage) {
     // TODO: Grey for invalid coverage
@@ -170,6 +171,7 @@ public class BadgeUtility {
    * Generates output file based on the extension give in configuration.
    * It expects renderedText to filled in with the actual data, coverage percentage
    * and label of the badge.
+   *
    * @param outputFile output file object pointing to badge location
    * @param renderedText rendered svg template having all the required data
    * @throws IOException if it's unable to perform file operations for the output.
