+++ src/main/java/io/github/handofgod94/format/Formatter.java
@@ -1,4 +1,4 @@
-package org.gahan.format;
+package io.github.handofgod94.format;
 
 import java.io.File;
 import java.io.IOException;
