+++ src/main/java/io/github/handofgod94/format/JpegFormatter.java
@@ -1,4 +1,4 @@
-package org.gahan.format;
+package io.github.handofgod94.format;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -8,16 +8,19 @@ import java.io.StringReader;
 import org.apache.batik.transcoder.TranscoderException;
 import org.apache.batik.transcoder.TranscoderInput;
 import org.apache.batik.transcoder.TranscoderOutput;
-import org.apache.batik.transcoder.image.PNGTranscoder;
+import org.apache.batik.transcoder.image.JPEGTranscoder;
 
 /**
- * Converts rendered SVG text to transparent PNG and saves it.
+ * Converts rendered SVG text to JPG and write it to output.
  */
-public class PngFormatter implements Formatter {
+public class JpegFormatter implements Formatter {
+
+  public static final float JPEG_QUALITY = 1.0f;
 
   @Override
   public void save(File file, String text) throws IOException, TranscoderException {
-    PNGTranscoder transcoder = new PNGTranscoder();
+    JPEGTranscoder transcoder = new JPEGTranscoder();
+    transcoder.addTranscodingHint(JPEGTranscoder.KEY_QUALITY, JPEG_QUALITY);
 
     TranscoderInput input = new TranscoderInput(new StringReader(text));
     OutputStream outputStream = new FileOutputStream(file);
