+++ src/main/resources/io/github/handofgod94/templates/svg-badge-template.ftl
